requirements.txt
OpenAi
Langchaink